requirements.txt
# Sistema de Transcrição de Áudio WhatsApp - Dependências Essenciais

# === CORE DEPENDENCIES ===
# FastAPI para API
fastapi==0.104.1
uvicorn[standard]==0.24.0

# === AUDIO PROCESSING ===
# Whisper para transcrição de áudio
openai-whisper

# PyTorch (CPU version - compatível com todos os sistemas)
# Para GPU, instale manualmente: pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu121
torch
torchaudio

# Processamento de áudio
ffmpeg-python
librosa
soundfile

# === DATABASE ===
# MongoDB
pymongo
motor

# === HTTP REQUESTS ===
requests
httpx

# === LLM INTEGRATION ===
# Ollama client (para análise de conversas)
ollama

# === UTILITIES ===
# Variáveis de ambiente
python-dotenv
pydantic
pydantic-settings
python-dateutil
orjson

# === MONITORING ===
psutil
prometheus-client

# === PRODUCTION ===
gunicorn

# === LOGGING ===
structlog
colorama